Moved helper methods to own file
